third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-2/lazy-computed-style-expected.txt

Tests that computed styles expand and allow tracing to style rules.

==== All styles (should be no computed) ====
[expanded] 
element.style { ()

[expanded] 
#inspected { (lazy-computed-style.html:15 -> lazy-computed-style.html:15:13)
    background: gray;
        background-image: initial;
        background-position-x: initial;
        background-position-y: initial;
        background-size: initial;
        background-repeat-x: initial;
        background-repeat-y: initial;
        background-attachment: initial;
        background-origin: initial;
        background-clip: initial;
        background-color: gray;

[expanded] 
#inspected { (lazy-computed-style.html:10 -> lazy-computed-style.html:10:13)
/-- overloaded --/     background-color: black;
    font-family: Courier;

[expanded] 
#inspected { (lazy-computed-style.html:5 -> lazy-computed-style.html:5:13)
/-- overloaded --/     background-color: green;
/-- overloaded --/     font-family: Times;

[expanded] 
div { (user agent stylesheet)
    display: block;

==== All styles (computed should be there) ====
background-attachment: scroll;
    initial - #inspected lazy-computed-style.html:15 -> lazy-computed-style.html:15:13
background-clip: border-box;
    initial - #inspected lazy-computed-style.html:15 -> lazy-computed-style.html:15:13
background-color: rgb(128, 128, 128);
    gray - #inspected lazy-computed-style.html:15 -> lazy-computed-style.html:15:13
    OVERLOADED black - #inspected lazy-computed-style.html:10 -> lazy-computed-style.html:10:13
    OVERLOADED green - #inspected lazy-computed-style.html:5 -> lazy-computed-style.html:5:13
background-image: none;
    initial - #inspected lazy-computed-style.html:15 -> lazy-computed-style.html:15:13
background-origin: padding-box;
    initial - #inspected lazy-computed-style.html:15 -> lazy-computed-style.html:15:13
background-position-x: 0%;
    initial - #inspected lazy-computed-style.html:15 -> lazy-computed-style.html:15:13
background-position-y: 0%;
    initial - #inspected lazy-computed-style.html:15 -> lazy-computed-style.html:15:13
background-repeat-x: ;
    initial - #inspected lazy-computed-style.html:15 -> lazy-computed-style.html:15:13
background-repeat-y: ;
    initial - #inspected lazy-computed-style.html:15 -> lazy-computed-style.html:15:13
background-size: auto;
    initial - #inspected lazy-computed-style.html:15 -> lazy-computed-style.html:15:13
display: block;
    block - div user agent stylesheet
font-family: Courier;
    Courier - #inspected lazy-computed-style.html:10 -> lazy-computed-style.html:10:13
    OVERLOADED Times - #inspected lazy-computed-style.html:5 -> lazy-computed-style.html:5:13
[expanded] 
element.style { ()

[expanded] 
#inspected { (lazy-computed-style.html:15 -> lazy-computed-style.html:15:13)
    background: gray;
        background-image: initial;
        background-position-x: initial;
        background-position-y: initial;
        background-size: initial;
        background-repeat-x: initial;
        background-repeat-y: initial;
        background-attachment: initial;
        background-origin: initial;
        background-clip: initial;
        background-color: gray;

[expanded] 
#inspected { (lazy-computed-style.html:10 -> lazy-computed-style.html:10:13)
/-- overloaded --/     background-color: black;
    font-family: Courier;

[expanded] 
#inspected { (lazy-computed-style.html:5 -> lazy-computed-style.html:5:13)
/-- overloaded --/     background-color: green;
/-- overloaded --/     font-family: Times;

[expanded] 
div { (user agent stylesheet)
    display: block;
